Add files via upload
The company that rents out the trailers work out the fee for the user by using a standard fee of 250 a day and works out if you should get a discount on your fee or not.

The discount is done by seeing if the trailer is used for more than 350km. If this is true, the total is deducted by 9% of the total. Else the total increases with 3%